Move query definitions into rustc_query_system.
