+++ compiler/rustc_query_system/src/lib.rs
@@ -13,6 +13,7 @@ extern crate rustc_data_structures;
 #[macro_use]
 extern crate rustc_macros;
 
+pub mod query_definitions;
 pub mod cache;
 pub mod dep_graph;
 pub mod query;
